+++ Sources/BugsnagPerformance/Private/BugsnagPerformancePriority+Private.h
@@ -1,0 +1,11 @@
+//
+//  BugsnagPerformancePriority+Private.h
+//  BugsnagPerformance
+//
+//  Created by Robert Bartoszewski on 25/05/2025.
+//  Copyright © 2025 Bugsnag. All rights reserved.
+//
+
+#import <BugsnagPerformance/BugsnagPerformancePriority.h>
+
+extern const BugsnagPerformancePriority BugsnagPerformancePriorityMax;