+++ Sources/BugsnagPerformance/Private/FrameRateMetrics/FrameMetricsSnapshot.mm
@@ -2,7 +2,7 @@
 //  FrameMetricsSnapshot.mm
 //  BugsnagPerformance
 //
-//  Created by Robert B on 23/08/2024.
+//  Created by Robert Bartoszewski on 23/08/2024.
 //  Copyright © 2024 Bugsnag. All rights reserved.
 //