+++ Sources/BugsnagPerformance/include/BugsnagPerformance/BugsnagPerformanceConfiguration.h
@@ -8,6 +8,7 @@
 #import <BugsnagPerformance/BugsnagPerformanceErrors.h>
 #import <BugsnagPerformance/BugsnagPerformanceNetworkRequestInfo.h>
 #import <BugsnagPerformance/BugsnagPerformanceSpan.h>
+#import <BugsnagPerformance/BugsnagPerformancePlugin.h>
 
 #import <UIKit/UIKit.h>
 
@@ -15,6 +16,15 @@ NS_ASSUME_NONNULL_BEGIN
 
 typedef BOOL (^ BugsnagPerformanceViewControllerInstrumentationCallback)(UIViewController *viewController);
 
+/**
+ * A callback that gets called whenever a span starts.
+ */
+typedef void (^ BugsnagPerformanceSpanStartCallback)(BugsnagPerformanceSpan *span);
+
+/**
+ * A callback that gets called whenever a span ends.
+ * @return If any of the registered callbacks returns false, the span is discarded.
+ */
 typedef BOOL (^ BugsnagPerformanceSpanEndCallback)(BugsnagPerformanceSpan *span);
 
 OBJC_EXPORT
@@ -40,10 +50,24 @@ OBJC_EXPORT
 - (BOOL) validate:(NSError * __autoreleasing _Nullable *)error NS_SWIFT_NAME(validate());
 
 /**
- * Add a callback that get called whenever a span ends.
+ * Add a callback that gets called whenever a span is started.
+ * This callback can be used to setup the span before it is
+ * used (such as setting attributes). These callbacks are considered to be "inside" the span, so
+ * any time taken to run the callback is counted towards the span's duration.
+ *
+ */
+- (void)addOnSpanStartCallback:(BugsnagPerformanceSpanStartCallback)callback;
+
+/**
+ * Add a callback that gets called whenever a span ends.
  * If any of the registered callbacks returns false, the span is discarded.
  */
-- (void) addOnSpanEndCallback:(BugsnagPerformanceSpanEndCallback) callback;
+- (void)addOnSpanEndCallback:(BugsnagPerformanceSpanEndCallback)callback;
+
+/**
+ * Add a plugin that gets called when the library is started.
+ */
+- (void)addPlugin:(id<BugsnagPerformancePlugin>)plugin;
 
 @property (nonatomic) NSString *apiKey;