+++ Sources/BugsnagPerformanceSwift/BugsnagPerformanceTrackedViewController.swift
@@ -2,7 +2,7 @@
 //  BugsnagPerformanceTrackedViewController.swift
 //  BugsnagPerformance-iOS
 //
-//  Created by Robert B on 07/12/2023.
+//  Created by Robert Bartoszewski on 07/12/2023.
 //  Copyright © 2023 Bugsnag. All rights reserved.
 //