+++ Tests/BugsnagPerformanceTests/SpanStackingHandlerTests.mm
@@ -2,7 +2,7 @@
 //  SpanStackingHandlerTests.mm
 //  BugsnagPerformance-iOSTests
 //
-//  Created by Robert B on 25/05/2023.
+//  Created by Robert Bartoszewski on 25/05/2023.
 //  Copyright © 2023 Bugsnag. All rights reserved.
 //